Update dependent classes for double

+++ src/main/java/seedu/address/model/person/Person.java
@@ -29,7 +29,7 @@ public class Person {
     private final Set<Tag> tags = new HashSet<>();
 
     private final List<Transaction> transactions;
-    private int balance;
+    private double balance;
 
     /**
      * Every field must be present and not null.
@@ -58,7 +58,7 @@ public class Person {
         this.address = address;
         this.tags.addAll(tags);
         this.transactions = new ArrayList<>(transactions);
-        this.balance = transactions.stream().map(Transaction::getAmount).reduce(0, Integer::sum);
+        this.balance = transactions.stream().map(Transaction::getAmount).reduce(0.0, Double::sum);
     }
 
     public Name getName() {
@@ -97,7 +97,7 @@ public class Person {
         this.transactions.remove(transaction);
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 
@@ -106,7 +106,7 @@ public class Person {
      *
      * @param amount the amount the balance is supposed to change by.
      */
-    public void updateBalance(int amount) {
+    public void updateBalance(double amount) {
         balance += amount;
     }
